float, not int

+++ app.py
@@ -535,12 +535,12 @@ def refresh_sabnzbd():
     sab_queue = json.loads(resp.text)
     print(sab_queue)
     log.info(sab_queue)
-    sabnzbd['status'] = sab_queue['queue']['status']
-    sabnzbd['queue_size'] = sab_queue['queue']['noofslots']
+    sabnzbd['status']           = sab_queue['queue']['status']
+    sabnzbd['queue_size']       = sab_queue['queue']['noofslots']
     sabnzbd['queue_space_left'] = sab_queue['queue']['sizeleft']
     sabnzbd['queue_time_left']  = sab_queue['queue']['timeleft']
     sabnzbd['free_space']       = sab_queue['queue']['diskspace1_norm']
-    sabnzbd['total_space']      = str(int(sab_queue['queue']['diskspacetotal1']) / 1000) + 'T'
+    sabnzbd['total_space']      = str(float(sab_queue['queue']['diskspacetotal1']) / 1000) + 'T'
 
 @app.route('/')
 def hello():
